Fix auCDtect.exe preventing removal of WAV Files
If a user invoked SIGINT when running '--aucdtect[-spectrogram]',
auCDtect.exe was not killed and would still process a WAV file
in the background.  This prevented removal of any WAV files to
to the dreated 'file or resource busy' message.

SoX is now able to handle files with unsupported characters.  Instead
of SoX handling the original FLAC file, it now handles the temporary
WAV file, which is ASCII named