Standardize spectrogram image names
This commit allows redoflacs to create the a spectrogram with
the same name as the filename, only this time, it's prepended
with [${iteration}], preventing any clobbering, even if all
the images are placed in a user-defined directory.

Previously, if a user had specified a user-defined directory
for the spectrograms, any file that was the same as another,
would have overwritten any image with the same name.  This is
no longer of consequence.

Also, the log file for auCDtect now uses the word 'REPORT' instead
of 'ERROR' to help the user understand that an error hasn't
occurred, only a report of auCDtect's findings