New option '--new-config, -o' and numerous fixes
New option '--new-config, -o' allows the user to force the
creation of a new configuration file.  Useful in case the
user b0rks his or hers.

This is also necessary as the script now checks the REVISION
number of the config file, checking if the config file the user
has is older than the one provided by redoflacs.  If that is
the case, the script will warn the user and give instructions
on how to update the configuration file.

There was a small grammar update to the configuration file
including it's first REVISION number (1).

Fixed issue with small terminal windows and not correctly
culling long filenames.

New and old warnings (that were missed) are now sent to
STDERR.

Documentation and man updates for the new option as well